+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v0.2.0]
  - bumped bascule, webpa-common, and wrp-go
  - removed temporary `/iot` endpoint 
  - updated release pipeline to use travis [#135](https://github.com/xmidt-org/tr1d1um/pull/135)
+ - added configurable way to check capabilities and put results into metrics, without rejecting requests [#137](https://github.com/xmidt-org/tr1d1um/pull/137)
 
 ## [v0.1.5]
  - migrated from glide to go modules
@@ -23,7 +26,8 @@ Switching to new build process
 ### Added
 - Initial creation
 
-[Unreleased]: https://github.com/xmidt-org/tr1d1um/compare/v0.1.5...HEAD
+[Unreleased]: https://github.com/xmidt-org/tr1d1um/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/xmidt-org/tr1d1um/compare/v0.1.5...v0.2.0
 [v0.1.5]: https://github.com/xmidt-org/tr1d1um/compare/v0.1.4...v0.1.5
 [v0.1.4]: https://github.com/xmidt-org/tr1d1um/compare/v0.1.2...v0.1.4
 [v0.1.2]: https://github.com/xmidt-org/tr1d1um/compare/0.1.1...v0.1.2